# docs/backend-api-contract.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Darts Backend API",
    "description": "OpenAPI-Dokumentation f\u00fcr Endpunkte unter `/api/*`.\nInteraktive Ansicht unter `/api/doc`, JSON-Schema unter `/api/doc.json`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/register": {
      "post": {
        "tags": ["Registration"],
        "summary": "Registers a user from JSON payload.",
        "operationId": "post_app_register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["username", "email", "plainPassword"],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "alice"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "alice@example.com"
                  },
                  "plainPassword": {
                    "type": "string",
                    "format": "password",
                    "example": "secret"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registrierung erfolgreich.",
            "content": {
              "application/json": {
                "schema": {
                  "required": ["success", "message", "status"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Registrierung erfolgreich"
                    },
                    "redirect": {
                      "type": "string",
                      "example": "/",
                      "nullable": true
                    },
                    "status": {
                      "type": "integer",
                      "example": 201
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validierungsfehler.",
            "content": {
              "application/json": {
                "schema": {
                  "required": ["success", "message", "status"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/login": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Returns login state or last error for API clients.",
        "operationId": "get_app_login",
        "requestBody": {
          "description": "Login-Daten f\u00fcr POST-Requests (Standard Symfony-Form-Felder). _username erwartet die E-Mail-Adresse.",
          "required": false,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "_username": {
                    "type": "string",
                    "example": "alice@example.com"
                  },
                  "_password": {
                    "type": "string",
                    "example": "secret"
                  },
                  "_csrf_token": {
                    "type": "string",
                    "example": "csrf-token"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login-Status. Wenn authentifiziert, enth\u00e4lt die Antwort eine Weiterleitung zu `/api/login/success`.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "id": {
                      "type": "integer",
                      "example": 1,
                      "nullable": true
                    },
                    "email": {
                      "type": "string",
                      "example": "alice@example.com",
                      "nullable": true
                    },
                    "username": {
                      "type": "string",
                      "example": "alice",
                      "nullable": true
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["ROLE_PLAYER"]
                    },
                    "redirect": {
                      "type": "string",
                      "example": "/api/login/success",
                      "nullable": true
                    },
                    "last_username": {
                      "type": "string",
                      "example": "alice",
                      "nullable": true
                    },
                    "error": {
                      "type": "string",
                      "example": "Ung\u00fcltige Zugangsdaten.",
                      "nullable": true
                    },
                    "status": {
                      "type": "integer",
                      "example": 401,
                      "nullable": true
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": []
      },
      "post": {
        "tags": ["Authentication"],
        "summary": "Returns login state or last error for API clients.",
        "operationId": "post_app_login",
        "requestBody": {
          "description": "Login-Daten f\u00fcr POST-Requests (Standard Symfony-Form-Felder). _username erwartet die E-Mail-Adresse.",
          "required": false,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "_username": {
                    "type": "string",
                    "example": "alice@example.com"
                  },
                  "_password": {
                    "type": "string",
                    "example": "secret"
                  },
                  "_csrf_token": {
                    "type": "string",
                    "example": "csrf-token"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login-Status. Wenn authentifiziert, enth\u00e4lt die Antwort eine Weiterleitung zu `/api/login/success`.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "id": {
                      "type": "integer",
                      "example": 1,
                      "nullable": true
                    },
                    "email": {
                      "type": "string",
                      "example": "alice@example.com",
                      "nullable": true
                    },
                    "username": {
                      "type": "string",
                      "example": "alice",
                      "nullable": true
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["ROLE_PLAYER"]
                    },
                    "redirect": {
                      "type": "string",
                      "example": "/api/login/success",
                      "nullable": true
                    },
                    "last_username": {
                      "type": "string",
                      "example": "alice",
                      "nullable": true
                    },
                    "error": {
                      "type": "string",
                      "example": "Ung\u00fcltige Zugangsdaten.",
                      "nullable": true
                    },
                    "status": {
                      "type": "integer",
                      "example": 401,
                      "nullable": true
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "name": "PHPSESSID",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "SessionCookie": []
    }
  ],
...
}